Rename module to create-settings
