+++ script/build.js
@@ -23,7 +23,7 @@ const watch = require('metalsmith-watch');
 const webpack = require('metalsmith-webpack');
 const webpackConfigGenerator = require('../config/webpack.config');
 const webpackDevServer = require('metalsmith-webpack-dev-server');
-const createSettings = require('../config/settings');
+const createSettings = require('../config/create-settings');
 
 const sourceDir = '../content/pages';
